Implementation up to PDF generation
